LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/animation/custom-property-transition-image-expected.txt


FAIL A custom property of type <image> can yield a CSS Transition assert_equals: A single animation is running expected 1 but got 0
